Fall_Detection/requirements.txt
matplotlib
torch
torchvision
opencv-python
xmltodict
fastapi
uvicorn[standard]
python-multipart
websockets
supabase
python-dotenv
scipy
